+++ examples/simple_pipelayer/src/app/app.py
@@ -1,16 +1,21 @@
+# region Imports
+# ==========================================================================
+# Uncomment this patch if running the example within of the pipelayer project
+# import app._path_patch  # NOQA F401
+
+import json
 from logging import Logger
 from typing import Any, Tuple
 
 from pipelayer import Pipeline
 
-# =========================================================================
-# Remove this patch if running the example outside of the pipelayer project
-import app._path_patch  # NOQA F401
 from app.app_context import AppContext
 from app.app_settings import AppSettings
+from app.filter.create_response import create_response
 from app.filter.hello_filter import HelloFilter
 from app.filter.world_filter import WorldFilter
-from app.render import print_manifest
+
+# endregion
 
 
 def main() -> Tuple[Pipeline, Any]:
@@ -18,14 +23,11 @@ def main() -> Tuple[Pipeline, Any]:
 
     context = AppContext(app_settings, Logger("Logger"))
     pipeline = Pipeline.create(context, "Hello World Pipeline")
-    filters = [HelloFilter(), WorldFilter()]
+    filters = [
+        HelloFilter,
+        WorldFilter("Append ', World.'", post_process=create_response),
+        lambda context, data: json.dumps(data)
+    ]
     output = pipeline.run(filters, None)
 
-    print_manifest(pipeline.manifest)
-
     return pipeline, output
-
-
-if __name__ == "__main__":
-    from app.render import print_output
-    print_output(main())