+++ examples/simple_pipelayer/src/app/filter/create_response.py
@@ -1,0 +1,5 @@
+from app.app_context import AppContext
+
+
+def create_response(context: AppContext, data: str) -> dict:
+    return {"message": data}