+++ examples/simple_pipelayer/src/app/filter/world_filter.py
@@ -4,6 +4,5 @@ from app.app_context import AppContext
 
 
 class WorldFilter(Filter):
-
     def run(self, context: AppContext, data: str) -> str:
-        return f"{data} World"
+        return f"{data}, World."