+++ src/pipelayer/__init__.py
@@ -1,10 +1,11 @@
-
 from __future__ import annotations
 
+import inspect
 from datetime import datetime
-from typing import Any, Callable, List, Optional
+from typing import Any, Callable, List, Optional, Tuple, Type, Union
 
 from pipelayer.context import Context
+from pipelayer.exception import InvalidFilterException, PipelineException
 from pipelayer.filter import Filter
 from pipelayer.final import Final
 from pipelayer.manifest import FilterManifestEntry, Manifest, ManifestEntry
@@ -12,9 +13,8 @@ from pipelayer.settings import Settings  # NOQA F401
 
 
 class Pipeline:
-
     __name: str
-    __context: Context
+    __context: Union[Context, Any]
     __manifest: Manifest
 
     @property
@@ -22,7 +22,7 @@ class Pipeline:
         return self.__name
 
     @property
-    def context(self) -> Context:
+    def context(self) -> Union[Context, Any]:
         return self.__context
 
     @property
@@ -30,7 +30,7 @@ class Pipeline:
         return self.__manifest
 
     @classmethod
-    def create(cls, context: Context, name: str = "") -> Pipeline:
+    def create(cls, context: Union[Context, Any], name: str = "") -> Pipeline:
         """
         Pipeline factory
         """
@@ -39,23 +39,34 @@ class Pipeline:
         pipeline.__context = context
         return pipeline
 
-    def run(self, filters: List[Filter], data: Any = None) -> Any:
+    def run(self, filters: List[Union[Filter, Callable[[Context, Any], Any]]], data: Any = None) -> Any:
         """
         The Pipeline runner
         """
         self.__manifest = Manifest(name=self.name, start=datetime.utcnow())
 
-        for filter in filters:
-            manifest_entry = FilterManifestEntry(name=filter.name, start=datetime.utcnow())
+        for filter in [Pipeline.__initialize_filter(filter) for filter in filters]:
+            filter_name, filter_func, pre_process, post_process = filter
+
+            manifest_entry = FilterManifestEntry(name=filter_name, start=datetime.utcnow())
 
             # pre_process
-            data = Pipeline.__run_filter_process(self.context, manifest_entry, filter.pre_process, data)
+            try:
+                manifest_entry.pre_process, data = self.__run_filter_process(pre_process, data)
+            except Exception as e:
+                raise PipelineException(inner_exception=e)
 
             # filter
-            data = filter.run(self.context, data)
+            try:
+                data = filter_func(self.context, data)
+            except Exception as e:
+                raise PipelineException(inner_exception=e)
 
             # post_process
-            data = Pipeline.__run_filter_process(self.context, manifest_entry, filter.post_process, data, False)
+            try:
+                manifest_entry.post_process, data = self.__run_filter_process(post_process, data)
+            except Exception as e:
+                raise PipelineException(inner_exception=e)
 
             manifest_entry.end = datetime.utcnow()
             manifest_entry.duration = (manifest_entry.end - manifest_entry.start)
@@ -69,25 +80,48 @@ class Pipeline:
         return data
 
     @staticmethod
-    def __run_filter_process(context: Context, filter_manifest_entry: FilterManifestEntry,
-                             process: Optional[Callable], data: Any, pre_process: bool = True) -> Any:
+    def __initialize_filter(filter: Union[Filter, Type[Filter], Callable[[Context, Any], Any]]) -> Tuple[
+        str,
+        Callable[[Context, Any], Any],
+        Optional[Callable[[Context, Any], Any]],
+        Optional[Callable[[Context, Any], Any]]
+    ]:
+        name = ""
+        if inspect.isclass(filter):
+            # The checks should have isolated the type, but mypy complains
+            filter = filter(filter.__name__)  # type: ignore
+
+        if isinstance(filter, Filter) and issubclass(filter.__class__, Filter):
+            return filter.name, filter.run, filter.pre_process, filter.post_process
+
+        if not Pipeline.__is_callable_valid(filter):
+            raise InvalidFilterException(("Filter functions cannot be 'None' and must have two arguments."))
+
+        name = f"[{inspect.getsource(filter).strip()}]" if filter.__name__ == "<lambda>" else filter.__name__
+
+        return name, filter, None, None
+
+    @staticmethod
+    def __is_callable_valid(obj: Callable[..., Any]) -> bool:
+        if not obj or not inspect.isfunction(obj):
+            return False
+        sig = inspect.signature(obj)
+        return len(sig.parameters) == 2
+
+    def __run_filter_process(self, process: Optional[Callable], data: Any) -> Tuple[Optional[ManifestEntry], Any]:
         """
         The filter pre/post process runner
         """
         if not process:
-            return data
+            return None, data
 
         process_manifest_entry = ManifestEntry(name=process.__name__, start=datetime.utcnow())
 
-        data = process(context, data)
+        data = process(self.context, data)
 
         process_manifest_entry.end = datetime.utcnow()
         process_manifest_entry.duration = (process_manifest_entry.end - process_manifest_entry.start)
 
-        if pre_process:
-            filter_manifest_entry.pre_process = process_manifest_entry
-        else:
-            filter_manifest_entry.post_process = process_manifest_entry
-        return data
+        return process_manifest_entry, data
 
     __metaclass__ = Final