+++ test/unit/conftest.py
@@ -1,7 +1,9 @@
 from test.fixtures.app_context import AppContext
 from test.fixtures.app_settings import AppSettings
+from test.fixtures.manifest import manifest_dict
 
 import pytest
+from pipelayer import Manifest
 
 
 @pytest.fixture
@@ -12,3 +14,8 @@ def app_settings() -> AppSettings:
 @pytest.fixture
 def app_context(app_settings) -> AppContext:
     return AppContext(app_settings)
+
+
+@pytest.fixture
+def manifest() -> dict:
+    return Manifest.parse_obj(manifest_dict)